+++ R/word_cor.R
@@ -104,7 +104,7 @@
 #' require(tm)
 #' data("crude")
 #' 
-#' out1 <- word_cor(t(tm_corpus2wfm(crude)), word = "oil", r=.7)
+#' out1 <- word_cor(t(as.wfm(crude)), word = "oil", r=.7)
 #' vect2df(out1[[1]], "word", "cor")
 #' 
 #' plot(out1)
@@ -112,7 +112,7 @@
 #'     digits=2, order.by ="cor", plot=FALSE) + coord_flip()
 #' 
 #' 
-#' out2 <- word_cor(t(tm_corpus2wfm(crude)), word = c("oil", "country"), r=.7)
+#' out2 <- word_cor(t(as.wfm(crude)), word = c("oil", "country"), r=.7)
 #' plot(out2)
 #' }
 word_cor <- function(text.var, grouping.var = NULL, word, r = .7,